+++ src/App.js
@@ -1,7 +1,13 @@
 import { useState, useEffect, React } from "react";
+import {
+  createArtistRecord,
+  bandReleases,
+  memberReleases,
+  contributorReleases,
+} from "./helpers/asyncCalls";
 
 function App() {
-  const [data, setData] = useState({ pagination: [], results: [], roles: [] });
+  const [data, setData] = useState([]);
   const [displayResults, setDisplayResults] = useState([]);
   const [excludeArtist, setExcludeArtist] = useState(true);
   const [formData, setFormData] = useState({
@@ -16,265 +22,49 @@ function App() {
 
   // console.log(process.env.SEARCH_URL);
 
-  const getSearchResult = async () => {
-    return await fetch(
-      `https://api.discogs.com/database/search?release_title=${album}&artist=${band}&type=release&sort=year&sort_order=asc&key=owJjvljKmrcdSbXFVPTu&secret=wgJurrmQFbROAyrmByuLrZMRMhDznPaK`
-    ).then((res) => res.json());
-  };
-
-  const bandReleases = async () => {
-    // searching with input params
-    const response = await getSearchResult();
-    console.log(response);
-    console.log(response.results);
-
-    // fetch the release information for the first result
-    const release = await fetch(response.results[0].resource_url).then((res) =>
-      res.json()
-    );
-    console.log(release);
-
-    // fetch the list of artists
-    const artists = await Promise.all(
-      release.artists.map(async (artist) => {
-        return await fetch(artist.resource_url).then((res) => res.json());
-      })
-    );
-    console.log(artists);
-
-    // for each artist, return their releases
-    const output = {
-      results: [],
-      pagination: [],
-      roles: [],
-    };
-    await Promise.all(
-      [...artists.values()].map(async (artist) => {
-        console.log(artist);
-        console.log(artist.id);
-        console.log(artist.roles);
-        const artistReleases = await fetch(
-          `https://api.discogs.com/artists/${artist.id}/releases?page=1&per_page=100`
-        )
-          .then((res) => res.json())
-          .catch((err) => console.log(err));
-
-        console.log(artistReleases);
-        if (
-          artistReleases &&
-          artistReleases.releases &&
-          artistReleases.pagination
-        ) {
-          output.pagination.push({
-            prev: artistReleases.pagination.urls?.prev,
-            next: artistReleases.pagination.urls?.next,
-          });
-          output.results.push(artistReleases.releases);
-          output.roles.push(artist.roles);
-        }
-      })
-    );
-    console.log(output);
-    return output;
-  };
-
-  const memberReleases = async (band) => {
-    console.log("searching for band...");
-
-    // searching with input params
-    const response = await getSearchResult();
-    console.log(response);
-    console.log(response.results);
-
-    // fetch the release information for the first result
-    const release = await fetch(response.results[0].resource_url).then((res) =>
-      res.json()
-    );
-    console.log(release);
-
-    // fetch the list of artists
-    const artists = await Promise.all(
-      release.artists.map(async (artist) => {
-        return await fetch(artist.resource_url).then((res) => res.json());
-      })
-    );
-    console.log(artists);
-
-    // for each artist get any members
-    const allMembers = [];
-
-    await Promise.all(
-      artists.map(async (artist) => {
-        if (artist.hasOwnProperty("members")) {
-          console.log(`${artist.name} has members`);
-          return await Promise.all(
-            artist.members.map(async (member) => {
-              const response = await fetch(member.resource_url).then((res) =>
-                res.json()
-              );
-              allMembers.push(response);
-            })
-          );
-        } else {
-          console.log(`${artist.name} has NO members`);
-          allMembers.push(artist);
-        }
-      })
-    );
-    console.log(allMembers);
-
-    // for each artist, return their releases
-    const output = {
-      results: [],
-      pagination: [],
-      roles: [],
-    };
-    await Promise.all(
-      [...allMembers.values()].map(async (member) => {
-        console.log(member);
-        console.log(member.id);
-        console.log(member.roles);
-        const memberReleases = await fetch(
-          `https://api.discogs.com/artists/${member.id}/releases?page=1&per_page=100`
-        )
-          .then((res) => res.json())
-          .catch((err) => console.log(err));
-
-        console.log(memberReleases);
-        if (
-          memberReleases &&
-          memberReleases.releases &&
-          memberReleases.pagination
-        ) {
-          output.pagination.push({
-            prev: memberReleases.pagination.urls?.prev,
-            next: memberReleases.pagination.urls?.next,
-          });
-          output.results.push(memberReleases.releases);
-          output.roles.push(member.roles);
-        }
-      })
-    );
-    console.log(output);
-    return output;
-  };
-
-  const contributorReleases = async (band, album) => {
-    console.log("searching for album...");
-    // searching with input params
-    const response = await getSearchResult();
-    console.log(response);
-    console.log(response.results);
-
-    // fetch the release information for the first result that has an extra artists
-    let release;
-    for (let i = 0; i < 5; i++) {
-      const nextRelease = await fetch(response.results[i].resource_url).then(
-        (res) => res.json()
-      );
-      console.log(i, nextRelease);
-      if (nextRelease.extraartists && nextRelease.extraartists.length > 0) {
-        release = nextRelease;
-        break;
-      }
-      release = nextRelease;
-    }
-    console.log(release);
-    console.log(release.extraartists);
-
-    // form a collection of relevant contributors and their roles
-    const contributors = new Map();
-    release.extraartists.forEach((extraArtist) => {
-      if (
-        !extraArtist.role.includes("Management") &&
-        !extraArtist.role.includes("Photo") &&
-        !extraArtist.role.includes("Translated") &&
-        !extraArtist.role.includes("Art") &&
-        !extraArtist.role.includes("Master")
-      ) {
-        const oldRoles = contributors.has(extraArtist.id)
-          ? [...contributors.get(extraArtist.id).roles]
-          : [];
-        const newRoles = extraArtist.role
-          .split(",")
-          .map((element) => element.trim());
-        contributors.set(extraArtist.id, {
-          id: extraArtist.id,
-          name: extraArtist.name,
-          link: extraArtist.resource_url,
-          roles: [...oldRoles, ...newRoles],
-        });
-      }
-    });
-    console.log(contributors);
-
-    // for each artist, return their releases
-    const output = {
-      results: [],
-      pagination: [],
-      roles: [],
-    };
-    await Promise.all(
-      [...contributors.values()].map(async (contributor) => {
-        console.log(contributor);
-        console.log(contributor.id);
-        console.log(contributor.roles);
-        const contributorReleases = await fetch(
-          `https://api.discogs.com/artists/${contributor.id}/releases?page=1&per_page=100`
-        )
-          .then((res) => res.json())
-          .catch((err) => console.log(err));
-
-        console.log(contributorReleases);
-        if (
-          contributorReleases &&
-          contributorReleases.releases &&
-          contributorReleases.pagination
-        ) {
-          output.pagination.push({
-            prev: contributorReleases.pagination.urls?.prev,
-            next: contributorReleases.pagination.urls?.next,
-          });
-          output.results.push(contributorReleases.releases);
-          output.roles.push(contributor.roles);
-        }
-      })
-    );
-    console.log(output);
-    return output;
-  };
-
   const loadMore = async (relation) => {
-    const searchURLs = [];
-    const output = {
-      results: [],
-      pagination: [],
-      roles: [],
-    };
-
-    data.pagination.forEach((links) => {
-      console.log(links[relation]);
-      if (links[relation] !== undefined) {
-        searchURLs.push(links[relation]);
-      }
-    });
-
+    const output = [];
     await Promise.all(
-      searchURLs.map(async (link) => {
-        const artistReleases = await fetch(link)
-          .then((res) => res.json())
-          .catch((err) => console.log(err.json()));
-
-        if (artistReleases.releases && artistReleases.pagination) {
-          output.pagination.push({
-            prev: artistReleases.pagination.urls?.prev,
-            next: artistReleases.pagination.urls?.next,
+      data.map(async (artist) => {
+        console.log(artist);
+        if (artist.pagination[relation] === undefined) {
+          output.push({
+            ...artist,
+            pagination: {
+              ...artist.pagination,
+              prev: artist.pagination.last,
+              last: artist.pagination.last,
+            },
+            releases: [],
           });
-          output.results.push(artistReleases.releases);
-          output.roles.push(artistReleases.roles);
+        } else {
+          const artistReleases = await fetch(artist.pagination[relation])
+            .then((res) => res.json())
+            .catch((err) => console.log(err));
+          if (
+            artistReleases &&
+            artistReleases.releases &&
+            artistReleases.pagination
+          ) {
+            console.log(artistReleases.pagination.urls);
+            const newArtist = createArtistRecord(
+              artist.name,
+              artist.id,
+              {
+                prev: artistReleases.pagination.urls?.prev,
+                next: artistReleases.pagination.urls?.next,
+                last: artist.pagination.last,
+              },
+              artistReleases.releases,
+              artist.roles
+            );
+            output.push(newArtist);
+          }
         }
       })
     );
+
+    console.log(output);
     return output;
   };
 
@@ -291,7 +81,7 @@ function App() {
   };
 
   const getBandReleases = async () => {
-    const releases = await bandReleases();
+    const releases = await bandReleases(band, album);
     console.log(releases);
     setData(releases);
     // setResults(releases.sort((a, b) => b.id - a.id));
@@ -309,12 +99,12 @@ function App() {
   };
 
   const getMemberReleases = async () => {
-    const releases = await memberReleases();
+    const releases = await memberReleases(band, album);
     setData(releases);
   };
 
   const getContributorReleases = async () => {
-    const releases = await contributorReleases();
+    const releases = await contributorReleases(band, album);
     setData(releases);
   };
 
@@ -338,8 +128,8 @@ function App() {
     //   }`;
     // });
     const display = [];
-    data.results.forEach((artistReleases, index) => {
-      artistReleases.map((release) => {
+    data.forEach((artist, index) => {
+      artist.releases.map((release) => {
         display.push(`${release.artist} - ${release.title} `);
       });
     });
@@ -402,13 +192,13 @@ function App() {
           ))}
       </div>
       <button
-        disabled={data.pagination.every((link) => link.prev === undefined)}
+        disabled={data.every((artist) => artist.pagination.prev === undefined)}
         onClick={loadLast}
       >
         Load Last Page
       </button>
       <button
-        disabled={data.pagination.every((link) => link.next === undefined)}
+        disabled={data.every((artist) => artist.pagination.next === undefined)}
         onClick={loadNext}
       >
         Load Next Page
